+++ src/execution/expansion_utils.c
@@ -6,7 +6,7 @@
 /*   By: ladloff <ladloff@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 12:16:53 by ladloff           #+#    #+#             */
-/*   Updated: 2024/04/08 12:05:57 by ladloff          ###   ########.fr       */
+/*   Updated: 2024/11/05 18:52:37 by ladloff          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,7 +82,7 @@ char	*create_new_string_with_value(t_master *master, char *str,
 		free(exp->value);
 		error_exit(master, "malloc (create_new_string_with_value)");
 	}
-	ft_strlcpy(new_str, str, exp->substr_start - str + 1);
+	ft_strlcpy(new_str, str, (size_t)(exp->substr_start - str + 1));
 	ft_strlcat(new_str, exp->value, len);
 	ft_strlcat(new_str, exp->substr_start + ft_strlen(exp->name) + 1, len);
 	return (new_str);
